web-portal-api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
PyJWT[crypto]>=2.8.0
boto3>=1.34.0
pydantic>=2.5.0
python-multipart>=0.0.6
websockets>=12.0
pydantic-settings>=2.0.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

